+++ src/screens/BookedScreen/index.tsx
@@ -1,12 +1,9 @@
-import React, { useCallback } from 'react';
-import { StyleSheet, View, FlatList } from 'react-native';
-
-import THEME from 'theme';
-import DATA from 'data';
-import Post from 'components/Post';
-import { IPost, AppNavigationParamList } from 'interfaces';
-
+import React from 'react';
 import { StackNavigationProp } from '@react-navigation/stack';
+
+import PostList from 'components/PostList';
+import { IPost, AppNavigationParamList } from 'interfaces';
+import DATA from 'data';
 
 type BookedScreenProps = {
   navigation: StackNavigationProp<AppNavigationParamList, 'BookedScreen'>;
@@ -22,34 +19,9 @@ const BookedScreen: React.FC<BookedScreenProps> = ({ navigation }) => {
     });
   };
 
-  const keyExtractor = useCallback((post: IPost) => post.id.toString(), []);
-
   const bookedData = DATA.filter((post: IPost) => post.booked === true);
 
-  return (
-    <View style={styles.container}>
-      <FlatList
-        data={bookedData}
-        keyExtractor={keyExtractor}
-        renderItem={({ item }) => <Post post={item} openItem={openItem} />}
-      />
-    </View>
-  );
+  return <PostList data={bookedData} openItem={openItem} />;
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 10,
-  },
-  title: {
-    fontSize: 24,
-    color: THEME.TEXT_COLOR,
-    fontFamily: 'OpenSans-Bold',
-  },
-  button: {
-    marginVertical: 10,
-  },
-});
 
 export default BookedScreen;
