+++ src/screens/MainScreen/index.tsx
@@ -1,12 +1,9 @@
-import React, { useCallback } from 'react';
-import { StyleSheet, View, FlatList } from 'react-native';
-
-import THEME from 'theme';
-import DATA from 'data';
-import Post from 'components/Post';
-import { IPost, AppNavigationParamList } from 'interfaces';
-
+import React from 'react';
 import { StackNavigationProp } from '@react-navigation/stack';
+
+import DATA from 'data';
+import { IPost, AppNavigationParamList } from 'interfaces';
+import PostList from 'components/PostList';
 
 type MainScreenProps = {
   navigation: StackNavigationProp<AppNavigationParamList, 'MainScreen'>;
@@ -22,32 +19,7 @@ const MainScreen: React.FC<MainScreenProps> = ({ navigation }) => {
     });
   };
 
-  const keyExtractor = useCallback((post: IPost) => post.id.toString(), []);
-
-  return (
-    <View style={styles.container}>
-      <FlatList
-        data={DATA}
-        keyExtractor={keyExtractor}
-        renderItem={({ item }) => <Post post={item} openItem={openItem} />}
-      />
-    </View>
-  );
+  return <PostList data={DATA} openItem={openItem} />;
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 10,
-  },
-  title: {
-    fontSize: 24,
-    color: THEME.TEXT_COLOR,
-    fontFamily: 'OpenSans-Bold',
-  },
-  button: {
-    marginVertical: 10,
-  },
-});
 
 export default MainScreen;
